Move the shared file to test classes.
